add index.html
It is the mgerged version of main page '/'.
Some of the js files are conflicting so further discussion is required
between Chanoon and James.
nav-bar.js is not included since the index.html is loading header.js,
which functions exactly same as nav-bar.js. I merged James' nav-bar
format and some of the functions but not everything yet.